Cleanup: move env fixtures to pytest_pootle.plugin
